Update php.php
Instead of running the code on our servers, this update uses a third-party API (rextester.com) to run the code. This API not only sanitizes the code, but all user-submitted code is run on their servers.This means that even if there is malicious code, it will not be on our end.